# cdc-cms-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CMS Service for CDC Integration - Schema change approval, table registry, mapping rules management.",
        "title": "CDC CMS Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/airbyte/import/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports selected streams from Airbyte into CMS Registry and creates default mapping rules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airbyte"
                ],
                "summary": "Execute smart import",
                "parameters": [
                    {
                        "description": "Import configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ImportStream"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/airbyte/import/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Shows streams from Airbyte that are not yet in the CMS Registry",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airbyte"
                ],
                "summary": "List importable streams",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/api/airbyte/jobs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the 20 most recent sync jobs from Airbyte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airbyte"
                ],
                "summary": "List Airbyte Jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cdc-cms-service_pkgs_airbyte.JobStatus"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/airbyte/sources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches all configured source connectors from Airbyte",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airbyte"
                ],
                "summary": "List Airbyte Sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cdc-cms-service_pkgs_airbyte.Source"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/airbyte/sync-audit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Compares Airbyte connection state with CMS Registry to find inconsistencies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airbyte"
                ],
                "summary": "Audit Sync State",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cdc-cms-service_pkgs_airbyte.StreamAudit"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/introspection/scan-raw/{table}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Directly scans the _raw_data column in the DW table to find fields not yet mapped. Works without Airbyte API.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Introspection"
                ],
                "summary": "Scan _raw_data JSONB for unmapped fields",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target table name (e.g. cdc_merchants)",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/introspection/scan/{table}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Scans the _raw_data column of a target table to find fields that are not yet mapped",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Introspection"
                ],
                "summary": "Scan raw data for new fields",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target table name",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cdc-cms-service_pkgs_airbyte.JobStatus": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "jobId": {
                    "type": "integer"
                },
                "status": {
                    "description": "E.g., \"running\", \"succeeded\", \"failed\"",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "cdc-cms-service_pkgs_airbyte.MismatchType": {
            "type": "string",
            "enum": [
                "status_mismatch",
                "missing_stream",
                "orphan_stream"
            ],
            "x-enum-comments": {
                "MismatchMissing": "e.g. Table in CMS but NOT in Airbyte connection",
                "MismatchOrphan": "e.g. Stream selected in Airbyte but NOT in CMS Registry",
                "MismatchStatus": "e.g. CMS active, Airbyte selected = false"
            },
            "x-enum-descriptions": [
                "e.g. CMS active, Airbyte selected = false",
                "e.g. Table in CMS but NOT in Airbyte connection",
                "e.g. Stream selected in Airbyte but NOT in CMS Registry"
            ],
            "x-enum-varnames": [
                "MismatchStatus",
                "MismatchMissing",
                "MismatchOrphan"
            ]
        },
        "cdc-cms-service_pkgs_airbyte.Source": {
            "type": "object",
            "properties": {
                "connectionId": {
                    "type": "string"
                },
                "database": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "string"
                },
                "sourceName": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "string"
                }
            }
        },
        "cdc-cms-service_pkgs_airbyte.StreamAudit": {
            "type": "object",
            "properties": {
                "airbyte_selected": {
                    "type": "boolean"
                },
                "mismatch_type": {
                    "$ref": "#/definitions/cdc-cms-service_pkgs_airbyte.MismatchType"
                },
                "stream_name": {
                    "type": "string"
                },
                "system_active": {
                    "type": "boolean"
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
        "internal_api.ImportStream": {
            "type": "object",
            "properties": {
                "connection_id": {
                    "type": "string"
                },
                "stream_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authenticate",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}